+++ Sources/Dragula/Dragula.swift
@@ -465,29 +465,3 @@ fileprivate struct DragulaSectionDropDelegate<Section: DragulaSection>: DropDele
     }
 }
 #endif
-
-/// An environment key to customize the corner radius of the drag preview.
-private struct DragPreviewCornerRadiusKey: EnvironmentKey {
-    static let defaultValue: CGFloat = 12
-}
-
-/// A SwiftUI environment value that controls the corner radius of draggable preview content.
-/// You can set this value on any view using `.environment(\.dragPreviewCornerRadius, radius)`
-extension EnvironmentValues {
-    /// The corner radius applied to drag preview content.
-    public var dragPreviewCornerRadius: CGFloat {
-        get { self[DragPreviewCornerRadiusKey.self] }
-        set { self[DragPreviewCornerRadiusKey.self] = newValue }
-    }
-}
-
-fileprivate extension View {
-    @ViewBuilder
-    func hidden(_ isHidden: Bool) -> some View {
-        if isHidden {
-            self.hidden()
-        } else {
-            self
-        }
-    }
-}
