+++ Sources/Dragula/Example/Section.swift
@@ -1,0 +1,24 @@
+//
+//  Section.swift
+//  Dragula
+//
+//  Created by Eugene Kovs on 04.07.2025.
+//  https://github.com/kovs705
+//  Contributed to Dragula
+//
+
+import SwiftUI
+
+struct Section: DragulaSection {
+        
+        struct Item: DragulaItem {
+            let id: UUID = .init()
+            let title: String
+            let color: Color
+            var isDraggable: Bool = true
+        }
+        
+        let id: UUID = .init()
+        let title: String
+        var items: [Item] = []
+    }
